fully commiting to lazy for neovim
