+++ home-manager/packages/personal.nix
@@ -4,6 +4,7 @@ with pkgs; [
     ansible
     doppler
     feh
+    fontforge
     pulumi-bin
     xclip
 ]
