+++ nvim/lua/harry/lazy.lua
@@ -1,126 +1,16 @@
 -- This file can be loaded by calling `lua require('plugins')` from your init.vim
 
-local lazypath = vim.fn.stdpath("data") .. "/lazy/lazy.nvim"
+local lazypath = vim.fn.stdpath('data') .. '/lazy/lazy.nvim'
 if not vim.loop.fs_stat(lazypath) then
     vim.fn.system({
-        "git",
-        "clone",
-        "--filter=blob:none",
-        "https://github.com/folke/lazy.nvim.git",
-        "--branch=stable",
+        'git',
+        'clone',
+        '--filter=blob:none',
+        'https://github.com/folke/lazy.nvim.git',
+        '--branch=stable',
         lazypath,
     })
 end
 vim.opt.rtp:prepend(lazypath)
 
-return require('lazy').setup({
-    {
-        'nvim-telescope/telescope.nvim',
-        tag = '0.1.1',
-        -- or                            , branch = '0.1.x',
-        dependencies = { { 'nvim-lua/plenary.nvim' }, { 'kdheepak/lazygit.nvim' } },
-        config = function()
-            require("telescope").load_extension("lazygit")
-        end,
-    },
-
-    {
-        'nvim-lualine/lualine.nvim',
-        dependencies = { 'kyazdani42/nvim-web-devicons' }
-    },
-
-    {
-        "catppuccin/nvim",
-        name = "catppuccin",
-        priority = 1000,
-        config = function()
-            require("catppuccin").setup({
-                flavour = "mocha", -- latte, frappe, macchiato, mocha
-                background = {
-                       -- :h background
-                    light = "latte",
-                    dark = "mocha",
-                },
-                transparent_background = true, -- disables setting the background color.
-                show_end_of_buffer = false, -- shows the '~' characters after the end of buffers
-                term_colors = false, -- sets terminal colors (e.g. `g:terminal_color_0`)
-                dim_inactive = {
-                    enabled = false, -- dims the background color of inactive window
-                    shade = "dark",
-                    percentage = 0.15, -- percentage of the shade to apply to the inactive window
-                },
-                no_italic = true,  -- Force no italic
-                no_bold = false,    -- Force no bold
-                no_underline = false, -- Force no underline
-                styles = {
-                                    -- Handles the styles of general hi groups (see `:h highlight-args`):
-                    comments = { "italic" }, -- Change the style of comments
-                    conditionals = { "italic" },
-                    loops = {},
-                    functions = {},
-                    keywords = {},
-                    strings = {},
-                    variables = {},
-                    numbers = {},
-                    booleans = {},
-                    properties = {},
-                    types = {},
-                    operators = {},
-                },
-                color_overrides = {},
-                custom_highlights = {},
-                integrations = {
-                    gitsigns = true,
-                    nvimtree = true,
-                    telescope = true,
-                    harpoon = true,
-                    mason = true,
-                    treesitter_context = true,
-                    treesitter = true,
-                    -- For more plugins integrations please scroll down (https://github.com/catppuccin/nvim#integrations)
-                },
-            })
-
-            -- setup must be called before loading
-            vim.cmd.colorscheme "catppuccin"
-        end
-    },
-
-    {
-        'nvim-treesitter/nvim-treesitter',
-        build = ':TSUpdate'
-    },
-
-    'nvim-treesitter/nvim-treesitter-context',
-    'theprimeagen/harpoon',
-    'mbbill/undotree',
-    'voldikss/vim-floaterm',
-    'lewis6991/gitsigns.nvim',
-    'preservim/nerdcommenter',
-
-    {
-        'VonHeikemen/lsp-zero.nvim',
-        branch = 'v1.x',
-        dependencies = {
-            -- LSP Support
-            { 'neovim/nvim-lspconfig' },             -- Required
-            { 'williamboman/mason.nvim' },           -- Optional
-            { 'williamboman/mason-lspconfig.nvim' }, -- Optional
-
-            -- Autocompletion
-            { 'hrsh7th/nvim-cmp' },         -- Required
-            { 'hrsh7th/cmp-nvim-lsp' },     -- Required
-            { 'hrsh7th/cmp-buffer' },       -- Optional
-            { 'hrsh7th/cmp-path' },         -- Optional
-            { 'saadparwaiz1/cmp_luasnip' }, -- Optional
-            { 'hrsh7th/cmp-nvim-lua' },     -- Optional
-
-            -- Snippets
-            { 'L3MON4D3/LuaSnip' },             -- Required
-            { 'rafamadriz/friendly-snippets' }, -- Optional
-
-            -- Rust
-            { 'simrat39/rust-tools.nvim' }
-        }
-    },
-})
+return require('lazy').setup('harry.plugins')
