+++ nvim/lua/harry/plugins/lsp.lua
@@ -1,0 +1,106 @@
+return {
+    'VonHeikemen/lsp-zero.nvim',
+    branch = 'v1.x',
+    dependencies = {
+        -- LSP Support
+        { 'neovim/nvim-lspconfig' },             -- Required
+        { 'williamboman/mason.nvim' },           -- Optional
+        { 'williamboman/mason-lspconfig.nvim' }, -- Optional
+
+        -- Autocompletion
+        { 'hrsh7th/nvim-cmp' },         -- Required
+        { 'hrsh7th/cmp-nvim-lsp' },     -- Required
+        { 'hrsh7th/cmp-buffer' },       -- Optional
+        { 'hrsh7th/cmp-path' },         -- Optional
+        { 'saadparwaiz1/cmp_luasnip' }, -- Optional
+        { 'hrsh7th/cmp-nvim-lua' },     -- Optional
+
+        -- Snippets
+        { 'L3MON4D3/LuaSnip' },             -- Required
+        { 'rafamadriz/friendly-snippets' }, -- Optional
+
+        -- Rust
+        { 'simrat39/rust-tools.nvim' }
+    },
+    config = function()
+        local lsp = require('lsp-zero')
+        lsp.preset('recommended')
+
+        lsp.ensure_installed({
+            'tsserver',
+            'eslint',
+            'rust_analyzer',
+            'nil_ls',
+        })
+
+        -- Fix Undefined global 'vim'
+        lsp.configure('lua_ls', {
+            settings = {
+                Lua = {
+                    diagnostics = {
+                        globals = { 'vim' }
+                    },
+                    runtime = {
+                        version = 'LUAJIT',
+                    },
+                    telemetry = {
+                        enable = false,
+                    },
+                }
+            }
+        })
+
+        -- Fix Random Yaml Errors
+        lsp.configure('yamlls', {
+            settings = {
+                yaml = {
+                    keyOrdering = false,
+                }
+            }
+        })
+
+        local cmp = require('cmp')
+        local cmp_select = { behavior = cmp.SelectBehavior.Select }
+        local cmp_mappings = lsp.defaults.cmp_mappings({
+            ['<C-p>'] = cmp.mapping.select_prev_item(cmp_select),
+            ['<C-n>'] = cmp.mapping.select_next_item(cmp_select),
+            ['<C-y>'] = cmp.mapping.confirm({ select = true }),
+            ["<C-Space>"] = cmp.mapping.complete(),
+        })
+
+        lsp.set_preferences({
+            sign_icons = {
+                error = "E",
+                warn = "W",
+                hint = "H",
+                info = "I",
+            }
+        })
+
+        lsp.setup_nvim_cmp({
+            mapping = cmp_mappings
+        })
+
+        lsp.on_attach(function(client, bufnr)
+            local opts = { buffer = bufnr, remap = false }
+
+            vim.keymap.set("n", "gd", function() vim.lsp.buf.definition() end, opts)
+            vim.keymap.set("n", "K", function() vim.lsp.buf.hover() end, opts)
+            vim.keymap.set("n", "<leader>vws", function() vim.lsp.buf.workspace_symbol() end, opts)
+            vim.keymap.set("n", "<leader>vd", function() vim.diagnostic.open_float() end, opts)
+            vim.keymap.set("n", "[d", function() vim.diagnostic.goto_next() end, opts)
+            vim.keymap.set("n", "]d", function() vim.diagnostic.goto_prev() end, opts)
+            vim.keymap.set("n", "<leader>ca", function() vim.lsp.buf.code_action() end, opts)
+            vim.keymap.set("n", "<leader>rr", function() vim.lsp.buf.references() end, opts)
+            vim.keymap.set("n", "<leader>rn", function() vim.lsp.buf.rename() end, opts)
+            vim.keymap.set("i", "<C-h>", function() vim.lsp.buf.signature_help() end, opts)
+            vim.keymap.set("n", '<leader>ff', function() vim.lsp.buf.format { async = true } end, opts)
+        end)
+
+        lsp.setup()
+
+        vim.diagnostic.config({
+            virtual_text = true
+        })
+    end
+}
