+++ nvim/lua/harry/plugins/misc.lua
@@ -1,0 +1,4 @@
+return {
+    'nvim-treesitter/nvim-treesitter-context',
+    'preservim/nerdcommenter',
+}
